requirements.txt
# requirements.txt
tweepy==4.14.0
pyyaml==6.0.1
python-dotenv==1.0.0